refactor: stream_proxy dissolution stage 3 - mixin-split StreamProxy (iter34)
Split the StreamProxy manager class (1447 lines / 71 methods) into 7 cohesive mixin
modules (<500 NLOC each): stream_proxy_mgr_{lifecycle,sessions,prefetch,context,handoff,
probe,faststart}. Composed via MRO: class StreamProxy(_MgrLifecycleMixin, ...). __init__,
the get_proxy()/reset_proxy_singleton() singleton plumbing, and settings reads stay on
StreamProxy. stream_proxy.py NLOC 2,968 -> ~1,099.

Method set identical (70/71 + __init__, 0 drop/dup, verified at runtime via dir()).
@staticmethod decorators preserved; mixin methods reach module-level/@patched names via
`import resources.lib.stream_proxy as _sp` + `_sp.<name>` (call-time, preserves @patch).
get_proxy() singleton, server start/stop, session lifecycle, fallback push/update,
prepare handoff, and settings reads all preserved.

Gauge held FLAT at 24 (decomposition checkpoint; file-nloc clears once HlsProducer is also
split). Verified: 2138 unit + 540 stream_proxy/readahead + 39 service + integration (real
ffmpeg) pass, pylint 10/10, ruff+black clean, vermin 3.8, every new module <500 NLOC.
