+++ repo/plugin.video.nzbdav/resources/lib/stream_proxy.py
@@ -45,7 +45,10 @@ from concurrent.futures import (  # noqa: F401  pylint: disable=unused-import
 from http.client import HTTPException
 from http.server import BaseHTTPRequestHandler, HTTPServer
 from socketserver import ThreadingMixIn as _ThreadingMixIn
-from urllib.request import Request, urlopen
+from urllib.request import (  # noqa: F401  pylint: disable=unused-import
+    Request,
+    urlopen,
+)
 
 import xbmc
 
@@ -63,7 +66,7 @@ except ImportError:
 # install) we fall back gracefully to None, which prepare_stream treats as a
 # failed faststart parse.
 try:
-    from resources.lib.mp4_parser import (  # noqa: E402
+    from resources.lib.mp4_parser import (  # noqa: E402,F401  pylint: disable=unused-import
         RangeCache,
         build_faststart_layout,
         fetch_remote_mp4_layout,
@@ -73,12 +76,18 @@ except (ImportError, ModuleNotFoundError):
     build_faststart_layout = None  # type: ignore[assignment]
     fetch_remote_mp4_layout = None  # type: ignore[assignment]
 
-from resources.lib import telemetry
-from resources.lib.dv_source import probe_dolby_vision_source
+from resources.lib import (  # noqa: F401  pylint: disable=unused-import
+    telemetry,
+)
+from resources.lib.dv_source import (  # noqa: F401  pylint: disable=unused-import
+    probe_dolby_vision_source,
+)
 from resources.lib.http_util import (  # noqa: F401  pylint: disable=unused-import
     notify as _notify,
 )
-from resources.lib.http_util import redact_text as _redact_text
+from resources.lib.http_util import (  # noqa: F401  pylint: disable=unused-import
+    redact_text as _redact_text,
+)
 
 # Singleton proxy instance
 _proxy = None
@@ -1837,8 +1846,45 @@ class HlsProducer:
         )
 
 
-class StreamProxy:
-    """Local HTTP proxy server for nzbdav streams."""
+# ---------------------------------------------------------------------------
+# Stage-3 StreamProxy mixin split. StreamProxy's methods now live in cohesive
+# ``stream_proxy_mgr_*`` mixins, composed back onto the class below via MRO.
+# Imported here (after _StreamHandler / HlsProducer are defined) so the class
+# declaration can reference the mixin names. The mixins reach this module's
+# globals at call time via ``import resources.lib.stream_proxy as _sp`` so test
+# monkeypatches keep resolving. __init__, the singleton plumbing, and the
+# module-level get_proxy()/reset helpers stay in this file.
+# ---------------------------------------------------------------------------
+from resources.lib.stream_proxy_mgr_context import (  # noqa: E402
+    _MgrContextBuildMixin,
+)
+from resources.lib.stream_proxy_mgr_faststart import (  # noqa: E402
+    _MgrFaststartMixin,
+)
+from resources.lib.stream_proxy_mgr_handoff import _MgrPrepareMixin  # noqa: E402
+from resources.lib.stream_proxy_mgr_lifecycle import (  # noqa: E402
+    _MgrLifecycleMixin,
+)
+from resources.lib.stream_proxy_mgr_prefetch import _MgrPrefetchMixin  # noqa: E402
+from resources.lib.stream_proxy_mgr_probe import _MgrProbeMixin  # noqa: E402
+from resources.lib.stream_proxy_mgr_sessions import _MgrSessionsMixin  # noqa: E402
+
+
+class StreamProxy(  # pylint: disable=too-many-ancestors
+    _MgrLifecycleMixin,
+    _MgrSessionsMixin,
+    _MgrPrefetchMixin,
+    _MgrContextBuildMixin,
+    _MgrPrepareMixin,
+    _MgrProbeMixin,
+    _MgrFaststartMixin,
+):
+    """Local HTTP proxy server for nzbdav streams.
+
+    The implementation is composed from cohesive ``stream_proxy_mgr_*``
+    mixins (lifecycle, sessions, prefetch, context build, prepare/handoff,
+    probing, faststart); ``__init__`` and the singleton plumbing stay here.
+    """
 
     def __init__(self):
         self._server = None
@@ -1848,1834 +1894,6 @@ class StreamProxy:
         self._prepare_lock = threading.RLock()
         self.prepare_token = uuid.uuid4().hex
         self._ffmpeg_capabilities = None
-
-    def start(self):
-        """Start the proxy server on a random port."""
-        self._server = _ThreadedHTTPServer(("127.0.0.1", 0), _StreamHandler)
-        self._server.owner_proxy = self
-        self._server.prepare_token = self.prepare_token
-        self.port = self._server.server_address[1]
-        self._refresh_ffmpeg_capabilities()
-        self._thread = threading.Thread(target=self._server.serve_forever)
-        self._thread.daemon = True
-        self._thread.start()
-        xbmc.log(
-            "NZB-DAV: Stream proxy started on port {}".format(self.port),
-            xbmc.LOGINFO,
-        )
-
-    def is_alive(self):
-        """True iff the proxy's HTTP server thread is still serving.
-
-        Returns False when either:
-        - The proxy hasn't been started yet (``_thread`` is None).
-        - The serve_forever thread has exited for any reason —
-          normal stop(), or an unhandled exception in the socket
-          accept loop (rare but has happened historically on
-          memory-pressure paths).
-
-        The service loop polls this once a second and restarts the
-        proxy when it drops to False so a crashed listener doesn't
-        silently wedge every subsequent /prepare call.
-        """
-        thread = self._thread
-        if thread is None:
-            return False
-        return thread.is_alive()
-
-    def stop(self):
-        """Stop the proxy server.
-
-        ``TCPServer.shutdown()`` only signals ``serve_forever`` to exit;
-        it does NOT close the listening socket. Pair it with
-        ``server_close()`` so the socket file descriptor is released
-        immediately. Otherwise the listener lingers until Python's GC
-        runs, which surfaces as ``ResourceWarning: unclosed socket`` in
-        the test suite and (more importantly) holds the port across a
-        rapid stop/start cycle on the real service.
-        """
-        self.clear_sessions()
-        if self._server:
-            self._server.shutdown()
-            self._server.server_close()
-            self._server = None
-        if self._thread:
-            self._thread.join(timeout=5)
-            self._thread = None
-
-    def clear_sessions(self, wait_for_process=True):
-        """Tear down every registered session and kill its ffmpeg process.
-
-        Called from:
-        - stop() on service shutdown
-        - prepare_stream() on each new play, so a zombie ffmpeg from a
-          previous stream that Kodi abandoned without firing onPlayBackStopped
-          (e.g. DB-vacuum stall that freezes the decoder) doesn't keep
-          writing into a half-dead TCP socket forever
-        - NzbdavPlayer stop/end hooks for clean-stop cases
-        """
-        if not self._server:
-            return
-        with self._context_lock:
-            sessions = list(getattr(self._server, "stream_sessions", {}).values())
-            pending = list(
-                getattr(self._server, "pending_stream_contexts", {}).values()
-            )
-            self._server.stream_sessions = {}
-            self._server.pending_stream_contexts = {}
-            self._server.stream_context = None
-            self._server.active_ffmpeg = None
-        seen = set()
-        for ctx in sessions + pending:
-            marker = id(ctx)
-            if marker in seen:
-                continue
-            seen.add(marker)
-            self._cleanup_session_or_defer(ctx, wait_for_process=wait_for_process)
-
-    def cleanup_session_by_id(self, session_id):
-        """Tear down a single session by id, cleanup outside the lock.
-
-        Used by the /prepare write-failure path when the plugin client
-        disconnects before the response is delivered (e.g. 60 s urlopen
-        timeout firing during a slow tempfile-faststart remux). Without
-        this, the session lingers until the next ``prepare_stream`` call
-        runs ``clear_sessions``, or the 6-hour TTL prune fires —
-        meaning a tempfile from a give-up'd play could occupy disk for
-        hours. Closes TODO.md §H.2-H12.
-        """
-        if not self._server or not session_id:
-            return
-        with self._context_lock:
-            sessions = getattr(self._server, "stream_sessions", {})
-            ctx = sessions.pop(session_id, None)
-            if ctx is not None and getattr(self._server, "stream_context", None) is ctx:
-                self._server.stream_context = None
-        if ctx is not None:
-            self._cleanup_session_or_defer(ctx)
-
-    def _cleanup_session_or_defer(self, ctx, wait_for_process=True):
-        """Cleanup now, or wait until in-flight handlers release the ctx."""
-        with self._context_lock:
-            if ctx.get("_cleanup_started"):
-                return
-            if int(ctx.get("_active_handlers", 0) or 0) > 0:
-                ctx["_cleanup_pending"] = True
-                return
-            ctx["_cleanup_started"] = True
-        self._cleanup_session(ctx, wait_for_process=wait_for_process)
-
-    @staticmethod
-    def _try_faststart_layout(remote_url, content_length, auth_header):
-        """Attempt virtual moov-relocation for an MP4.
-
-        Returns the faststart dict, or None on failure.
-        """
-        try:
-            if fetch_remote_mp4_layout is None:
-                raise ImportError("mp4_parser not available")
-            layout_info = fetch_remote_mp4_layout(
-                remote_url, content_length, auth_header
-            )
-            if layout_info:
-                xbmc.log(
-                    "NZB-DAV: MP4 layout: moov_before_mdat={}, moov={}B".format(
-                        layout_info.get("moov_before_mdat"),
-                        len(layout_info.get("moov_data", b"")),
-                    ),
-                    xbmc.LOGINFO,
-                )
-                faststart = build_faststart_layout(layout_info)
-                if faststart is None:
-                    xbmc.log(
-                        "NZB-DAV: stco overflow — moov relocation failed "
-                        "(file >4GB with 32-bit chunk offsets)",
-                        xbmc.LOGWARNING,
-                    )
-                return faststart
-            xbmc.log("NZB-DAV: MP4 layout fetch returned None", xbmc.LOGWARNING)
-            return None
-        except _PARSE_ERRORS as e:
-            xbmc.log(
-                "NZB-DAV: MP4 faststart parse failed: {}".format(e), xbmc.LOGWARNING
-            )
-            return None
-
-    @staticmethod
-    def _wait_for_active_ffmpeg(active_ffmpeg):
-        try:
-            active_ffmpeg.wait(timeout=5)
-        except (OSError, subprocess.SubprocessError, ValueError):
-            pass
-
-    @staticmethod
-    def _wait_for_active_ffmpeg_in_background(active_ffmpeg):
-        thread = threading.Thread(
-            target=StreamProxy._wait_for_active_ffmpeg,
-            args=(active_ffmpeg,),
-            name="nzbdav-old-ffmpeg-reap",
-        )
-        thread.daemon = True
-        thread.start()
-
-    @staticmethod
-    def _cleanup_session(ctx, wait_for_process=True):
-        """Release resources associated with a stream session."""
-        # Signal the read-ahead prefetch daemon to stop (per-session close; the
-        # thread is daemon and also aborts on waitForAbort for global shutdown).
-        # This is the single sink reached by every close route, so all sessions
-        # are covered. Never block teardown on the thread — daemon is the
-        # backstop.
-        readahead_buffer = ctx.get(_READAHEAD_BUFFER_KEY)
-        if readahead_buffer is not None:
-            readahead_buffer.stop()
-
-        StreamProxy._kill_session_ffmpeg(ctx, wait_for_process)
-        StreamProxy._remove_session_tempfile(ctx)
-
-        hls_producer = ctx.get("hls_producer")
-        if hls_producer is not None:
-            try:
-                hls_producer.close(wait_for_process=wait_for_process)
-            except _HLS_CLOSE_ERRORS as e:
-                xbmc.log(
-                    "NZB-DAV: HLS producer close failed: {}".format(e),
-                    xbmc.LOGWARNING,
-                )
-
-    @staticmethod
-    def _kill_session_ffmpeg(ctx, wait_for_process):
-        """Kill the session's ffmpeg process and reap it (sync or background)."""
-        active_ffmpeg = ctx.get("active_ffmpeg")
-        if not active_ffmpeg:
-            return
-        try:
-            active_ffmpeg.kill()
-        except (OSError, subprocess.SubprocessError, ValueError):
-            return
-        if wait_for_process:
-            StreamProxy._wait_for_active_ffmpeg(active_ffmpeg)
-        else:
-            StreamProxy._wait_for_active_ffmpeg_in_background(active_ffmpeg)
-
-    @staticmethod
-    def _remove_session_tempfile(ctx):
-        """Best-effort removal of the session's temp faststart file."""
-        temp_path = ctx.get("temp_path")
-        if temp_path and os.path.exists(temp_path):
-            try:
-                os.remove(temp_path)
-            except OSError:
-                pass
-
-    @staticmethod
-    def _probe_hls_fmp4_capability(ffmpeg_path):
-        """Return True when ffmpeg exposes the HLS fMP4 muxer flags we use."""
-        if not ffmpeg_path:
-            return False
-        output = _run_ffmpeg_hls_muxer_probe(ffmpeg_path)
-        if output is None:
-            return False
-        supported = all(marker in output for marker in _FMP4_HLS_CAPABILITY_MARKERS)
-        xbmc.log(
-            "NZB-DAV: ffmpeg fmp4 HLS capability {} ({})".format(
-                "present" if supported else "absent", ffmpeg_path
-            ),
-            xbmc.LOGINFO if supported else xbmc.LOGWARNING,
-        )
-        return supported
-
-    def _refresh_ffmpeg_capabilities(self):
-        """Discover ffmpeg once so service-start logs show the active muxers."""
-        ffmpeg_path = _find_ffmpeg()
-        capabilities = {
-            "ffmpeg_path": ffmpeg_path,
-            "hls_fmp4": self._probe_hls_fmp4_capability(ffmpeg_path),
-        }
-        self._ffmpeg_capabilities = capabilities
-        return capabilities
-
-    def _get_ffmpeg_capabilities(self):
-        """Return cached ffmpeg capabilities, probing lazily if needed."""
-        capabilities = getattr(self, "_ffmpeg_capabilities", None)
-        if isinstance(capabilities, dict):
-            return capabilities
-        ffmpeg_path = _find_ffmpeg()
-        return {
-            "ffmpeg_path": ffmpeg_path,
-            "hls_fmp4": bool(ffmpeg_path),
-        }
-
-    def _assert_context_lock_owned(self):
-        """Best-effort debug guard for helpers that require _context_lock."""
-        if not __debug__:
-            return
-        context_lock = getattr(self, "_context_lock", None)
-        is_owned = getattr(context_lock, "_is_owned", None)
-        if callable(is_owned) and not is_owned():
-            raise AssertionError("_prune_sessions_locked requires _context_lock")
-
-    def _add_pending_context(self, session_id, ctx):
-        """Register ctx in the server's pending-context map (locked)."""
-        with self._context_lock:
-            pending = getattr(self._server, "pending_stream_contexts", None)
-            if not isinstance(pending, dict):
-                pending = {}
-                self._server.pending_stream_contexts = pending
-            pending[session_id] = ctx
-
-    def _drop_pending_context(self, session_id):
-        """Remove session_id from the server's pending-context map (locked)."""
-        with self._context_lock:
-            pending = getattr(self._server, "pending_stream_contexts", {})
-            if isinstance(pending, dict):
-                pending.pop(session_id, None)
-
-    @staticmethod
-    def _rewrite_ctx_to_matroska(ctx):
-        """Rewrite an HLS ctx in-place to the known-good matroska shape.
-
-        ctx already has ffmpeg_path / total_bytes / duration_seconds from
-        prepare_stream's fmp4 branch, so _serve_remux has everything it
-        needs once the HLS-specific keys are dropped.
-        """
-        ctx.pop("mode", None)
-        ctx.pop("hls_segment_format", None)
-        ctx.pop("hls_segment_duration", None)
-        ctx.pop("hls_producer", None)
-        ctx["content_type"] = "video/x-matroska"
-        ctx["seekable"] = (
-            ctx.get("duration_seconds") is not None and ctx.get("total_bytes", 0) > 0
-        )
-
-    def _register_hls_session(self, ctx, session_id):
-        """Spawn the HLS producer for ctx, falling back to matroska.
-
-        Eager spawn-time validation catches ffmpeg builds that reject
-        ``-hls_segment_type fmp4`` BEFORE the HLS URL is returned to Kodi,
-        so the matroska rewrite fires for the most likely failure mode
-        (no-op for mpegts, which spawns lazily). Raises RuntimeError if
-        the prepare was cancelled mid-flight.
-        """
-        self._add_pending_context(session_id, ctx)
-        workdir = _choose_hls_workdir(ctx.get("total_bytes", 0) or 0)
-        producer = None
-        try:
-            producer = HlsProducer(ctx, workdir)
-            ctx["hls_producer"] = producer
-            producer.prepare()
-        except Exception as e:  # noqa: BLE001 — fall back either way
-            if ctx.get("_cleanup_started"):
-                self._drop_pending_context(session_id)
-                raise RuntimeError("HLS prepare was cancelled") from e
-            xbmc.log(
-                "NZB-DAV: HLS producer setup failed ({}), "
-                "rewriting session to matroska fallback".format(e),
-                xbmc.LOGWARNING,
-            )
-            # Best-effort cleanup of the partially initialized producer.
-            # HlsProducer.__init__ owns disk resources (session_dir,
-            # ffmpeg.log) that need close()'ing on the prepare()-failure
-            # path; the `producer = None` sentinel protects against
-            # AttributeError when the constructor itself raised.
-            if producer is not None:
-                try:
-                    producer.close()
-                except Exception:  # noqa: BLE001
-                    pass
-            self._rewrite_ctx_to_matroska(ctx)
-        finally:
-            self._drop_pending_context(session_id)
-        if ctx.get("_cleanup_started"):
-            raise RuntimeError("HLS prepare was cancelled")
-
-    def _register_session(self, ctx):
-        """Store a per-stream context and return its unique proxy URL.
-
-        The returned URL shape depends on ``ctx["mode"]``:
-
-        - ``"hls"`` → ``/hls/<session>/playlist.m3u8`` so Kodi's HLS
-          demuxer takes over and drives segment fetches.
-        - default → ``/stream/<session>`` for the existing faststart /
-          temp-faststart / remux / pass-through handlers.
-
-        For HLS sessions, an ``HlsProducer`` is attached to the ctx
-        (``ctx["hls_producer"]``) which owns the persistent ffmpeg
-        process and the on-disk segment directory.
-        """
-        session_id = uuid.uuid4().hex
-        now = time.time()
-        ctx["session_id"] = session_id
-        ctx["created_at"] = now
-        ctx["last_access"] = now
-        ctx["ffmpeg_lock"] = threading.Lock()
-        ctx["active_ffmpeg"] = None
-        ctx["current_byte_pos"] = 0
-        ctx["_active_handlers"] = 0
-        ctx["_cleanup_pending"] = False
-        ctx["_cleanup_started"] = False
-
-        if ctx.get("mode") == "hls":
-            self._register_hls_session(ctx, session_id)
-
-        with self._context_lock:
-            if not isinstance(getattr(self._server, "stream_sessions", None), dict):
-                self._server.stream_sessions = {}
-            self._server.stream_context = ctx
-            self._server.stream_sessions[session_id] = ctx
-            evicted = self._prune_sessions_locked(keep_session=session_id)
-        # Cleanup outside the lock — `_cleanup_session` does proc.kill +
-        # proc.wait, which on a stuck ffmpeg can block other lock waiters.
-        for evicted_ctx in evicted:
-            self._cleanup_session_or_defer(evicted_ctx)
-
-        if ctx.get("mode") == "hls":
-            return "http://127.0.0.1:{}/hls/{}/playlist.m3u8".format(
-                self.port, session_id
-            )
-        return "http://127.0.0.1:{}/stream/{}".format(self.port, session_id)
-
-    def _prune_sessions_locked(self, keep_session=None):
-        """Drop expired sessions and cap the total number retained.
-
-        Returns the list of evicted ctx dicts so the *caller* can do the
-        actual `_cleanup_session(ctx)` work OUTSIDE `_context_lock`. The
-        cleanup path runs `proc.kill()` + `proc.wait()` and on a stuck
-        ffmpeg can block long enough to wedge any other thread waiting
-        on `_context_lock`. Closes TODO.md §H.2-H1e.
-        """
-        self._assert_context_lock_owned()
-        sessions = getattr(self._server, "stream_sessions", {})
-        now = time.time()
-        evicted = []
-
-        expired = _expired_session_ids(sessions, keep_session, now)
-        for session_id in expired:
-            ctx = sessions.pop(session_id, None)
-            if ctx is not None:
-                evicted.append(ctx)
-
-        while len(sessions) > _MAX_STREAM_SESSIONS:
-            session_id = _least_recently_used_session(sessions, keep_session)
-            if session_id is None:
-                break
-            ctx = sessions.pop(session_id, None)
-            if ctx is not None:
-                evicted.append(ctx)
-
-        return evicted
-
-    def _refresh_session_standby_fallbacks(self, ctx):
-        """Resolve nzo-only standby fallbacks into usable WebDAV stream URLs.
-
-        Runs the same resolution the failure-time cutover would, but AHEAD of
-        the failure, so a primary stall becomes an instant source swap instead
-        of a multi-second cold resolve under Kodi's read timeout.
-        """
-        handler = _StreamHandler.__new__(_StreamHandler)
-        handler.server = self._server
-        try:
-            handler._refresh_standby_fallback_sources(ctx)
-        except Exception as exc:  # pylint: disable=broad-except
-            xbmc.log(
-                "NZB-DAV: standby fallback resolve failed: {}".format(exc),
-                xbmc.LOGWARNING,
-            )
-
-    def _prevalidate_fallback_sources(self, ctx):
-        handler = _StreamHandler.__new__(_StreamHandler)
-        handler.server = self._server
-        try:
-            validated = handler._prevalidate_ready_fallback_sources(ctx)
-        except Exception as exc:  # pylint: disable=broad-except
-            xbmc.log(
-                "NZB-DAV: Fallback prevalidation failed: {}".format(exc),
-                xbmc.LOGWARNING,
-            )
-            return
-        if validated:
-            xbmc.log(
-                "NZB-DAV: Prevalidated {} fallback stream(s)".format(validated),
-                xbmc.LOGINFO,
-            )
-
-    def _start_fallback_prevalidation(self, ctx):
-        expected_length = _StreamHandler._fallback_expected_content_length(ctx)
-        if expected_length <= 0:
-            xbmc.log(
-                "NZB-DAV: prevalidation skipped (expected_length={})".format(
-                    expected_length
-                ),
-                xbmc.LOGINFO,
-            )
-            return
-        sources = ctx.get("fallback_sources") or []
-        pending = [s for s in sources if _fallback_source_needs_prevalidation(s)]
-        xbmc.log(
-            "NZB-DAV: prevalidation pending sources={}/{} expected_length={}".format(
-                len(pending), len(sources), expected_length
-            ),
-            xbmc.LOGINFO,
-        )
-        if not pending:
-            return
-
-        # Coalesce bursts: fallbacks are pushed one-per-adopted-job, so mark
-        # the session dirty and let a single running warmer re-loop to pick up
-        # late arrivals, instead of spawning a thread per push.
-        ctx["_fallback_prevalidation_dirty"] = True
-        if _thread_is_alive(ctx.get("_fallback_prevalidation_thread")):
-            return
-
-        def _warm():
-            self._run_fallback_prevalidation_warmer(ctx)
-
-        thread = threading.Thread(target=_warm, name="nzbdav-fallback-prevalidate")
-        thread.daemon = True
-        ctx["_fallback_prevalidation_thread"] = thread
-        try:
-            thread.start()
-        except RuntimeError:
-            # Out of thread budget — match the sibling prefetch spawns and fail
-            # soft so prevalidation never wedges /prepare.
-            ctx.pop("_fallback_prevalidation_thread", None)
-
-    def _run_fallback_prevalidation_warmer(self, ctx):
-        """Warmer-thread body: drain the dirty flag, resolving + fingerprinting.
-
-        Resolves nzo-only standbys into WebDAV URLs, then fingerprints the
-        resolved sources so the live cutover is an instant pointer swap. The
-        dirty flag lets a merge that lands mid-pass trigger one more pass; it
-        terminates once no new push has arrived.
-        """
-        prefetch_thread = ctx.get("_initial_range_prefetch_thread")
-        if prefetch_thread and prefetch_thread is not threading.current_thread():
-            try:
-                prefetch_thread.join()
-            except RuntimeError:
-                pass
-        while ctx.pop("_fallback_prevalidation_dirty", False):
-            self._refresh_session_standby_fallbacks(ctx)
-            self._prevalidate_fallback_sources(ctx)
-
-    @staticmethod
-    def _initial_range_prefetchable(ctx):
-        """Return whether this pass-through context can prefetch byte 0."""
-        if (
-            ctx.get("remux")
-            or ctx.get("faststart")
-            or ctx.get("temp_faststart")
-            or ctx.get("mode") == "hls"
-            or not ctx.get("remote_url")
-        ):
-            return False
-        try:
-            return int(ctx.get("content_length", 0) or 0) > 0
-        except (TypeError, ValueError):
-            return False
-
-    def _prefetch_initial_range(self, ctx):
-        """Prime the first pass-through bytes for Kodi's initial range GET."""
-        try:
-            content_length = int(ctx.get("content_length", 0) or 0)
-        except (TypeError, ValueError):
-            return
-        if content_length <= 0:
-            return
-        end = min(content_length - 1, _UPSTREAM_READ_CHUNK - 1)
-        try:
-            body = _StreamHandler._fetch_primary_range_bytes(
-                ctx["remote_url"],
-                ctx.get("auth_header"),
-                0,
-                end,
-                content_length,
-            )
-        except Exception as exc:  # pylint: disable=broad-except
-            xbmc.log(
-                "NZB-DAV: Initial range prefetch failed: {}".format(exc),
-                xbmc.LOGDEBUG,
-            )
-            return
-        if body:
-            _StreamHandler._cache_fallback_range(
-                ctx,
-                ctx["remote_url"],
-                ctx.get("auth_header"),
-                content_length,
-                0,
-                body,
-            )
-
-    def _start_initial_range_prefetch(self, ctx):
-        """Start byte-0 prefetch without delaying proxy prepare."""
-        if not self._initial_range_prefetchable(ctx):
-            return
-        thread = threading.Thread(
-            target=self._prefetch_initial_range,
-            args=(ctx,),
-            name="nzbdav-initial-range-prefetch",
-        )
-        thread.daemon = True
-        ctx["_initial_range_prefetch_thread"] = thread
-        try:
-            thread.start()
-        except RuntimeError:
-            ctx.pop("_initial_range_prefetch_thread", None)
-
-    def _run_readahead_prefetch(self, ctx):
-        """Daemon target: fill the read-ahead window forward of the play head.
-
-        Reads sequentially AHEAD of the highest-served offset in 64KB chunks
-        using the same upstream-fetch primitive the serve path uses, throttles
-        when the window is full (so it keeps filling WHILE PAUSED yet bounded),
-        and resumes as free-behind reclaims room. Best-effort: a prefetch
-        upstream error simply backs off and retries WITHOUT tearing down
-        playback or tripping the user-facing recovery taxonomy / notifications /
-        session counters. NEVER raises into anything.
-        """
-        setup = self._readahead_prefetch_setup(ctx)
-        if setup is None:
-            return
-        buf, monitor, content_length = setup
-        if self._readahead_defer_start(buf, monitor):
-            return
-        while not buf.should_stop() and not monitor.waitForAbort(0):
-            try:
-                if self._readahead_prefetch_once(ctx, buf, monitor, content_length):
-                    return
-            except Exception as exc:  # pylint: disable=broad-except
-                xbmc.log(
-                    "NZB-DAV: Read-ahead prefetch loop error: {}".format(exc),
-                    xbmc.LOGDEBUG,
-                )
-                if monitor.waitForAbort(_READAHEAD_ERROR_BACKOFF_SECONDS):
-                    return
-
-    @staticmethod
-    def _readahead_prefetch_setup(ctx):
-        """Validate ctx and build the read-ahead loop inputs.
-
-        Returns ``(buf, monitor, content_length)`` when prefetch should run,
-        or None when the ctx is missing a buffer / remote_url / positive
-        content_length (the daemon then exits without looping).
-        """
-        buf = ctx.get(_READAHEAD_BUFFER_KEY) if isinstance(ctx, dict) else None
-        if buf is None:
-            return None
-        try:
-            content_length = int(ctx.get("content_length", 0) or 0)
-        except (TypeError, ValueError):
-            return None
-        if not ctx.get("remote_url") or content_length <= 0:
-            return None
-        return buf, xbmc.Monitor(), content_length
-
-    @staticmethod
-    def _readahead_defer_start(buf, monitor):
-        """Yield to startup before the first read-ahead fetch; True to abort.
-
-        Lets the byte-0 prefetch and Kodi's first range fetch win nzbdav's
-        connection budget before the read-ahead issues its first upstream
-        read. Abortable so a shutdown during the defer exits cleanly. Returns
-        True when the loop should not start (stop requested or abort).
-        """
-        return buf.should_stop() or monitor.waitForAbort(_READAHEAD_START_DEFER_SECONDS)
-
-    @staticmethod
-    def _readahead_prefetch_once(ctx, buf, monitor, content_length):
-        """Run one read-ahead fetch iteration; return True to stop the loop.
-
-        Returns True when the lead has reached EOF or an abort was signalled
-        during a backoff wait (the loop should `return`); False to continue
-        to the next iteration. Mirrors the original inline body's throttle /
-        error-backoff branches exactly.
-        """
-        fetch_offset = buf.next_fetch_offset()
-        if fetch_offset >= content_length:
-            # Lead is fully built to EOF; nothing left to prefetch.
-            return True
-        if buf.is_full():
-            # Throttle: wait (abortably) for free-behind to reclaim room as
-            # the play head advances. This is what keeps the buffer filling
-            # while paused yet strictly bounded.
-            return monitor.waitForAbort(_READAHEAD_THROTTLE_BACKOFF_SECONDS)
-        want = min(
-            _READAHEAD_FETCH_CHUNK,
-            buf.space_remaining(),
-            content_length - fetch_offset,
-        )
-        if want <= 0:
-            return monitor.waitForAbort(_READAHEAD_THROTTLE_BACKOFF_SECONDS)
-        end = fetch_offset + want - 1
-        # Re-read the source each iteration: a live fallback cutover mutates
-        # ctx["remote_url"]/["auth_header"] mid-stream, so a once-hoisted local
-        # would keep hammering the dead primary and the lead would stop growing
-        # from the live source. The buffer is offset-addressed, so bytes from
-        # either source are interchangeable for a given offset.
-        remote_url = ctx.get("remote_url")
-        auth_header = ctx.get("auth_header")
-        if not remote_url:
-            return monitor.waitForAbort(_READAHEAD_ERROR_BACKOFF_SECONDS)
-        body = _StreamHandler._fetch_primary_range_bytes(
-            remote_url, auth_header, fetch_offset, end, content_length
-        )
-        if not body:
-            # Best-effort upstream error or awaiting-download: back off and
-            # retry. The real serve path owns recovery; do not touch the
-            # recovery taxonomy / notifications / counters here.
-            return monitor.waitForAbort(_READAHEAD_ERROR_BACKOFF_SECONDS)
-        buf.append(fetch_offset, body)
-        return False
-
-    def _start_readahead_prefetch(self, ctx):
-        """Spawn the per-session read-ahead prefetch daemon (gated + soft).
-
-        Gated on the same passthrough-only guard as the byte-0 prefetch AND on
-        readahead_buffer_mb>0 (read from the prefetched settings snapshot, never
-        a per-request getSetting). RuntimeError-tolerant start matching the
-        sibling prefetch spawns.
-        """
-        if not self._initial_range_prefetchable(ctx):
-            return
-        settings = _passthrough_runtime_settings(ctx)
-        mb = _coerce_nonneg_int(settings.get("readahead_buffer_mb", 0))
-        if mb <= 0:
-            return
-        content_length = _coerce_nonneg_int(ctx.get("content_length", 0))
-        if content_length <= 0:
-            return
-        cap_bytes = mb * 1024 * 1024
-        buf = ReadAheadBuffer(cap_bytes, content_length)
-        ctx[_READAHEAD_BUFFER_KEY] = buf
-        thread = threading.Thread(
-            target=self._run_readahead_prefetch,
-            args=(ctx,),
-            name="nzbdav-readahead",
-        )
-        thread.daemon = True
-        ctx[_READAHEAD_THREAD_KEY] = thread
-        try:
-            thread.start()
-        except RuntimeError:
-            ctx.pop(_READAHEAD_THREAD_KEY, None)
-            ctx.pop(_READAHEAD_BUFFER_KEY, None)
-
-    def _prewarm_tail_range(self, ctx):
-        """Warm nzbdav's FILE-TAIL article cache (MKV cues) for instant startup.
-
-        Kodi reads the MKV cues/SeekHead at the file tail before it can play. For
-        a usenet-backed file nzbdav fetches those end-of-file articles on demand,
-        so Kodi's first tail read otherwise stalls 1-4s mid-startup — long enough
-        to drain its not-yet-full cache and wedge the CoreELEC audio clock
-        (permanent black screen). Issue a throwaway read of the last
-        _TAIL_PREWARM_BYTES here, during the prepare gap, so nzbdav has the tail
-        articles cached before Kodi asks. No proxy-side caching: the exact tail
-        offset Kodi requests isn't fixed, and nzbdav serves subsequent tail reads
-        fast once the articles are fetched. The body is discarded.
-        """
-        try:
-            content_length = int(ctx.get("content_length", 0) or 0)
-        except (TypeError, ValueError):
-            return
-        # Only warm a tail that is distinct from the byte-0 prefetch window;
-        # tiny files are already fully covered by _prefetch_initial_range.
-        if content_length <= _TAIL_PREWARM_BYTES + _UPSTREAM_READ_CHUNK:
-            return
-        # Yield to startup playback: hold the tail read back a short, abortable
-        # beat so the byte-0 prefetch and Kodi's first-byte range request win
-        # nzbdav's connection budget first. Without this the tail read RACED
-        # them at prepare time and widened the very mid-startup stall window the
-        # prewarm exists to close (transient black-screen). waitForAbort lets a
-        # Kodi shutdown / session stop cancel the prewarm cleanly during the
-        # defer instead of blocking the daemon thread or wasting a connection.
-        if xbmc.Monitor().waitForAbort(_TAIL_PREWARM_DEFER_SECONDS):
-            return
-        tail_start = content_length - _TAIL_PREWARM_BYTES
-        try:
-            _StreamHandler._fetch_primary_range_bytes(
-                ctx["remote_url"],
-                ctx.get("auth_header"),
-                tail_start,
-                content_length - 1,
-                content_length,
-            )
-        except Exception as exc:  # pylint: disable=broad-except
-            xbmc.log(
-                "NZB-DAV: Tail prewarm failed: {}".format(exc),
-                xbmc.LOGDEBUG,
-            )
-
-    def _start_tail_prewarm(self, ctx):
-        """Warm the file tail in parallel with the byte-0 prefetch at prepare.
-
-        Runs on its own thread so it neither delays /prepare nor serializes
-        behind the byte-0 prefetch — Kodi reads the tail FIRST, so warming it
-        must start as early as possible. Fails soft when out of thread budget,
-        matching the sibling prefetch spawns.
-        """
-        if not self._initial_range_prefetchable(ctx):
-            return
-        thread = threading.Thread(
-            target=self._prewarm_tail_range,
-            args=(ctx,),
-            name="nzbdav-tail-prewarm",
-        )
-        thread.daemon = True
-        ctx["_tail_prewarm_thread"] = thread
-        try:
-            thread.start()
-        except RuntimeError:
-            ctx.pop("_tail_prewarm_thread", None)
-
-    def _start_passthrough_runtime_settings_prefetch(self, ctx):
-        """Read pass-through recovery settings during the player handoff gap."""
-        if isinstance(ctx.get(_PASSTHROUGH_RUNTIME_SETTINGS_KEY), dict):
-            return
-        if not self._initial_range_prefetchable(ctx):
-            return
-        done = threading.Event()
-        ctx[_PASSTHROUGH_RUNTIME_SETTINGS_DONE_KEY] = done
-
-        def _worker():
-            try:
-                ctx[_PASSTHROUGH_RUNTIME_SETTINGS_KEY] = (
-                    _read_passthrough_runtime_settings()
-                )
-            except Exception as exc:  # pylint: disable=broad-except
-                ctx[_PASSTHROUGH_RUNTIME_SETTINGS_ERROR_KEY] = exc
-            finally:
-                done.set()
-
-        thread = threading.Thread(
-            target=_worker,
-            name="nzbdav-passthrough-settings-prefetch",
-        )
-        thread.daemon = True
-        ctx["_passthrough_runtime_settings_thread"] = thread
-        try:
-            thread.start()
-        except RuntimeError:
-            ctx.pop("_passthrough_runtime_settings_thread", None)
-            ctx.pop(_PASSTHROUGH_RUNTIME_SETTINGS_DONE_KEY, None)
-
-    def merge_session_fallbacks(self, session_id, fallback_sources):
-        """Merge late-adopted fallback sources into a live session context.
-
-        Returns the count of NEW sources added, or None when the session is
-        unknown (torn down / never existed). Dedups by (nzo_id, stream_url).
-        Existing source dicts are preserved BY IDENTITY so in-place
-        failed/validated marks the live cutover writes survive a merge, and
-        the list is swapped atomically so a concurrent _serve_proxy reader
-        never observes a half-mutated list.
-        """
-        normalized = _normalize_fallback_sources(fallback_sources)
-
-        with self._context_lock:
-            sessions = getattr(self._server, "stream_sessions", None)
-            if not isinstance(sessions, dict):
-                return None
-            ctx = sessions.get(session_id)
-            if not isinstance(ctx, dict):
-                return None
-            existing = list(ctx.get("fallback_sources") or [])
-            added = _merge_new_fallback_sources(existing, normalized)
-            if added:
-                ctx["fallback_sources"] = existing
-        if added:
-            # Warm the freshly-pushed fallbacks in the background (resolve
-            # nzo-only standbys + fingerprint) so a later primary failure cuts
-            # over instantly instead of cold-resolving under Kodi's timeout.
-            self._start_fallback_prevalidation(ctx)
-        return added
-
-    @staticmethod
-    def _probe_dv_source(remote_url, auth_header, content_length):
-        """Run the pure-Python DV source probe, failing safe on crash."""
-        try:
-            dv_result = probe_dolby_vision_source(
-                remote_url,
-                auth_header,
-                file_size=content_length if content_length else None,
-            )
-        except Exception as exc:  # pylint: disable=broad-except
-            xbmc.log(
-                "NZB-DAV: DV probe crashed -- failing safe to "
-                "matroska: {!r}".format(exc),
-                xbmc.LOGWARNING,
-            )
-            from resources.lib.dv_source import DolbyVisionSourceResult
-
-            dv_result = DolbyVisionSourceResult("dv_unknown", "probe_crashed")
-        xbmc.log(
-            "NZB-DAV: dv_probe classification={} reason={} "
-            "profile={} el_type={}".format(
-                dv_result.classification,
-                dv_result.reason,
-                dv_result.profile,
-                dv_result.el_type,
-            ),
-            xbmc.LOGDEBUG,
-        )
-        return dv_result
-
-    def _dv_route_allows_fmp4(self, remote_url, auth_header, content_length):
-        """Gate fmp4 HLS on DV profile via the pure-Python source probe.
-
-        Returns True when the source may use fmp4 HLS, False when it must
-        fall back to matroska. Probes the first HEVC access unit to classify
-        DV profile/EL and applies the 2026-04-23 routing matrix verbatim.
-        """
-        dv_result = self._probe_dv_source(remote_url, auth_header, content_length)
-        if dv_result.classification == "dv_profile_7_fel":
-            xbmc.log(
-                "NZB-DAV: dv_route=matroska reason={} "
-                "profile={} el_type={}".format(
-                    dv_result.reason,
-                    dv_result.profile,
-                    dv_result.el_type,
-                ),
-                xbmc.LOGWARNING,
-            )
-            return False
-        if (
-            dv_result.classification == "dv_allowed_for_fmp4"
-            and dv_result.profile == 7
-            and dv_result.el_type == "MEL"
-        ):
-            xbmc.log(
-                "NZB-DAV: dv_route=fmp4 reason={} profile=7 "
-                "el_type=MEL (experimental -- metadata-only EL "
-                "does not exercise CAMLCodec dual-layer init)".format(dv_result.reason),
-                xbmc.LOGINFO,
-            )
-            return True
-        return self._dv_route_tail(dv_result)
-
-    @staticmethod
-    def _dv_route_tail(dv_result):
-        """Tail of the DV routing matrix: non-P7 fmp4, non_dv, unknown."""
-        if dv_result.classification == "dv_allowed_for_fmp4":
-            xbmc.log(
-                "NZB-DAV: dv_route=matroska reason={} "
-                "profile={} (non-P7 DV hangs CAMLCodec "
-                "onAVStarted on fmp4 per 2026-04-15 "
-                "testing)".format(dv_result.reason, dv_result.profile),
-                xbmc.LOGWARNING,
-            )
-            return False
-        if dv_result.classification == "non_dv":
-            xbmc.log(
-                "NZB-DAV: dv_route=fmp4 reason={}".format(dv_result.reason),
-                xbmc.LOGDEBUG,
-            )
-            return True
-        xbmc.log(
-            "NZB-DAV: dv_route=matroska reason={} "
-            "profile={} el_type={} (unknown DV state -- "
-            "failing safe)".format(
-                dv_result.reason,
-                dv_result.profile,
-                dv_result.el_type,
-            ),
-            xbmc.LOGWARNING,
-        )
-        return False
-
-    def _build_ctx_fallback(
-        self, remote_url, auth_header, content_length_hint, content_type, is_mp4
-    ):
-        """Pass-through proxy context for fallback-enabled streams."""
-        content_length = self._get_content_length(
-            remote_url, auth_header, content_length_hint=content_length_hint
-        )
-        if content_length <= 0:
-            raise OSError(
-                "Unable to determine content length for fallback-enabled stream"
-            )
-        ctx = {
-            "remote_url": remote_url,
-            "auth_header": auth_header,
-            "content_length": content_length,
-            "content_type": "video/mp4" if is_mp4 else content_type,
-            "remux": False,
-            "faststart": False,
-            "seekable": True,
-        }
-        xbmc.log(
-            "NZB-DAV: Fallback streams attached; using pass-through proxy "
-            "before MP4 repair or remux rescue tiers",
-            xbmc.LOGINFO,
-        )
-        return ctx
-
-    def _resolve_mp4_temp_path(
-        self,
-        ffmpeg_path,
-        remote_url,
-        auth_header,
-        content_length,
-        content_length_unknown,
-    ):
-        """Tier 2 decision: temp-file faststart path, or None to skip.
-
-        Skip for large files (>4GB) — temp remux would take too long and
-        would time out the prepare_stream_via_service call.
-        """
-        _TEMP_FASTSTART_MAX = 4 * 1073741824  # 4 GB
-        if content_length_unknown:
-            xbmc.log(
-                "NZB-DAV: MP4 content length unknown; skipping temp-file faststart",
-                xbmc.LOGWARNING,
-            )
-            return None
-        if content_length > _TEMP_FASTSTART_MAX:
-            xbmc.log(
-                "NZB-DAV: File too large for temp-file faststart "
-                "({}B > {}B), skipping to MKV remux".format(
-                    content_length, _TEMP_FASTSTART_MAX
-                ),
-                xbmc.LOGINFO,
-            )
-            return None
-        if ffmpeg_path:
-            return self._prepare_tempfile_faststart(
-                ffmpeg_path, remote_url, auth_header
-            )
-        return None
-
-    def _build_ctx_mp4_tempfile_or_remux(
-        self, remote_url, auth_header, content_length, content_length_unknown
-    ):
-        """MP4 tiers 2/3: temp-file faststart, MKV remux, or direct proxy."""
-        # Tier 2: Try temp-file faststart (ffmpeg -movflags +faststart)
-        ffmpeg_path = self._get_ffmpeg_capabilities().get("ffmpeg_path")
-        temp_path = self._resolve_mp4_temp_path(
-            ffmpeg_path,
-            remote_url,
-            auth_header,
-            content_length,
-            content_length_unknown,
-        )
-
-        if temp_path:
-            temp_size = os.path.getsize(temp_path)
-            ctx = {
-                "remote_url": remote_url,
-                "auth_header": auth_header,
-                "content_type": "video/mp4",
-                "faststart": False,
-                "remux": False,
-                "temp_faststart": True,
-                "temp_path": temp_path,
-                "content_length": temp_size,
-            }
-            xbmc.log(
-                "NZB-DAV: MP4 temp-file faststart ({}B)".format(temp_size),
-                xbmc.LOGINFO,
-            )
-            return ctx
-        if ffmpeg_path:
-            return self._ctx_mp4_mkv_remux(
-                remote_url, auth_header, ffmpeg_path, content_length
-            )
-        if content_length_unknown:
-            raise OSError(
-                "Unable to determine content length for MP4 stream "
-                "and ffmpeg unavailable"
-            )
-        # Last resort: direct proxy (may fail for large files)
-        return {
-            "remote_url": remote_url,
-            "auth_header": auth_header,
-            "content_length": content_length,
-            "content_type": "video/mp4",
-            "remux": False,
-            "faststart": False,
-        }
-
-    def _ctx_mp4_mkv_remux(self, remote_url, auth_header, ffmpeg_path, content_length):
-        """Tier 3: MKV remux fallback for an MP4 source (existing behavior)."""
-        duration = self._probe_duration(ffmpeg_path, remote_url, auth_header)
-        ctx = {
-            "remote_url": remote_url,
-            "auth_header": auth_header,
-            "content_type": "video/x-matroska",
-            "remux": True,
-            "faststart": False,
-            "ffmpeg_path": ffmpeg_path,
-            "total_bytes": content_length,
-            "duration_seconds": duration,
-            "seekable": duration is not None and content_length > 0,
-        }
-        xbmc.log("NZB-DAV: MP4 fallback to MKV remux", xbmc.LOGWARNING)
-        return ctx
-
-    def _build_ctx_mp4(self, remote_url, auth_header, content_length_hint):
-        """MP4 proxy context: faststart layout, then temp/remux tiers."""
-        content_length = self._get_content_length(
-            remote_url, auth_header, content_length_hint=content_length_hint
-        )
-        content_length_unknown = content_length <= 0
-        faststart = self._try_faststart_layout(remote_url, content_length, auth_header)
-
-        if faststart is not None and not faststart.get("already_faststart"):
-            ctx = {
-                "remote_url": remote_url,
-                "auth_header": auth_header,
-                "content_type": "video/mp4",
-                "faststart": True,
-                "remux": False,
-                "header_data": faststart["header_data"],
-                "virtual_size": faststart["virtual_size"],
-                "payload_remote_start": faststart["payload_remote_start"],
-                "payload_remote_end": faststart["payload_remote_end"],
-                "payload_size": faststart["payload_size"],
-                "range_cache": RangeCache(),
-            }
-            xbmc.log(
-                "NZB-DAV: MP4 faststart proxy (virtual={}B, header={}B)".format(
-                    faststart["virtual_size"], len(faststart["header_data"])
-                ),
-                xbmc.LOGINFO,
-            )
-            return ctx
-        if faststart is not None and faststart.get("already_faststart"):
-            xbmc.log(
-                "NZB-DAV: MP4 already faststart; using pass-through proxy",
-                xbmc.LOGINFO,
-            )
-            return {
-                "remote_url": remote_url,
-                "auth_header": auth_header,
-                "content_length": content_length,
-                "content_type": "video/mp4",
-                "remux": False,
-                "faststart": False,
-                "seekable": True,
-            }
-        return self._build_ctx_mp4_tempfile_or_remux(
-            remote_url, auth_header, content_length, content_length_unknown
-        )
-
-    def _build_ctx_default_remux(
-        self,
-        remote_url,
-        auth_header,
-        content_length,
-        force_mode,
-        ffmpeg_caps,
-        threshold,
-    ):
-        """ffmpeg-available remux context: fmp4 HLS or piped MKV.
-
-        Force-remux exists for 32-bit Kodi builds (Amlogic CoreELEC and
-        similar) that throw ``Open - Unhandled exception`` on pass-through
-        HTTP above ~4 GB Content-Length. force_remux_mode picks the shape:
-        "matroska" (default, piped MKV, cache-bounded seek) or "hls_fmp4"
-        (experimental fragmented-MP4 HLS VOD, full random seek, DV-capable).
-        """
-        ffmpeg_path = ffmpeg_caps.get("ffmpeg_path")
-        duration = self._probe_duration(ffmpeg_path, remote_url, auth_header)
-        use_fmp4 = (
-            force_mode == "hls_fmp4"
-            and ffmpeg_caps.get("hls_fmp4", False)
-            and duration is not None
-            and duration > 0
-        )
-        if force_mode == "hls_fmp4" and not ffmpeg_caps.get("hls_fmp4", False):
-            xbmc.log(
-                "NZB-DAV: ffmpeg lacks required fmp4 HLS flags; "
-                "falling back to piped Matroska",
-                xbmc.LOGWARNING,
-            )
-        if use_fmp4:
-            use_fmp4 = self._dv_route_allows_fmp4(
-                remote_url, auth_header, content_length
-            )
-        if use_fmp4:
-            return self._ctx_fmp4_hls(
-                remote_url, auth_header, ffmpeg_path, content_length, duration
-            )
-        return self._ctx_piped_mkv(
-            remote_url, auth_header, ffmpeg_path, content_length, duration, threshold
-        )
-
-    @staticmethod
-    def _ctx_fmp4_hls(remote_url, auth_header, ffmpeg_path, content_length, duration):
-        """fMP4 HLS remux context (experimental, full random seek)."""
-        ctx = {
-            "remote_url": remote_url,
-            "auth_header": auth_header,
-            "content_type": "application/vnd.apple.mpegurl",
-            "mode": "hls",
-            "remux": True,
-            "faststart": False,
-            "ffmpeg_path": ffmpeg_path,
-            "total_bytes": content_length,
-            "duration_seconds": duration,
-            "seekable": True,
-            "hls_segment_duration": _HLS_SEGMENT_SECONDS,
-            "hls_segment_format": "fmp4",
-        }
-        xbmc.log(
-            "NZB-DAV: Force-remuxing {}B file via fMP4 HLS "
-            "(experimental, duration={:.1f}s)".format(content_length, duration),
-            xbmc.LOGWARNING,
-        )
-        return ctx
-
-    @staticmethod
-    def _ctx_piped_mkv(
-        remote_url, auth_header, ffmpeg_path, content_length, duration, threshold
-    ):
-        """Piped Matroska remux context (cache-bounded seek)."""
-        ctx = {
-            "remote_url": remote_url,
-            "auth_header": auth_header,
-            "content_type": "video/x-matroska",
-            "remux": True,
-            "faststart": False,
-            "ffmpeg_path": ffmpeg_path,
-            "total_bytes": content_length,
-            "duration_seconds": duration,
-            "seekable": duration is not None and content_length > 0,
-        }
-        xbmc.log(
-            "NZB-DAV: Force-remuxing large {}B file via piped MKV "
-            "(duration={}, threshold={}B)".format(
-                content_length,
-                "{:.1f}s".format(duration) if duration else "unknown",
-                threshold,
-            ),
-            xbmc.LOGWARNING,
-        )
-        return ctx
-
-    @staticmethod
-    def _decide_force_remux(content_length, content_length_unknown, settings_snapshot):
-        """Resolve (threshold, force_mode, needs_remux) for a non-MP4 source."""
-        if settings_snapshot:
-            threshold = _force_remux_threshold_bytes_from_snapshot(settings_snapshot)
-            force_mode = _force_remux_mode_from_snapshot(settings_snapshot)
-        else:
-            threshold = _get_force_remux_threshold_bytes()
-            force_mode = _get_force_remux_mode()
-        force_remux_requested = threshold > 0 and force_mode in (
-            "matroska",
-            "hls_fmp4",
-        )
-        needs_remux = force_remux_requested and (
-            content_length_unknown or content_length >= threshold
-        )
-        if needs_remux and content_length_unknown:
-            xbmc.log(
-                "NZB-DAV: Content length unknown; forcing live remux "
-                "instead of zero-byte pass-through",
-                xbmc.LOGWARNING,
-            )
-        return threshold, force_mode, needs_remux
-
-    def _build_ctx_default(
-        self,
-        remote_url,
-        auth_header,
-        content_length_hint,
-        content_type,
-        settings_snapshot,
-    ):
-        """Non-MP4 context: force-remux decision then remux or pass-through."""
-        content_length = self._get_content_length(
-            remote_url, auth_header, content_length_hint=content_length_hint
-        )
-        content_length_unknown = content_length <= 0
-        threshold, force_mode, needs_remux = self._decide_force_remux(
-            content_length, content_length_unknown, settings_snapshot
-        )
-        ffmpeg_caps = self._get_ffmpeg_capabilities() if needs_remux else {}
-        if ffmpeg_caps.get("ffmpeg_path"):
-            return self._build_ctx_default_remux(
-                remote_url,
-                auth_header,
-                content_length,
-                force_mode,
-                ffmpeg_caps,
-                threshold,
-            )
-        if needs_remux:
-            if content_length_unknown:
-                raise OSError(
-                    "Unable to determine content length for stream "
-                    "and ffmpeg unavailable"
-                )
-            xbmc.log(
-                "NZB-DAV: {}B file exceeds remux threshold but no "
-                "ffmpeg found — falling back to pass-through, "
-                "playback may fail on 32-bit Kodi".format(content_length),
-                xbmc.LOGWARNING,
-            )
-        return {
-            "remote_url": remote_url,
-            "auth_header": auth_header,
-            "content_length": content_length,
-            "content_type": content_type,
-            "remux": False,
-        }
-
-    def _validate_and_classify(
-        self,
-        remote_url,
-        auth_header,
-        fallback_sources,
-        content_length_hint,
-        settings_snapshot,
-    ):
-        """Validate/normalize inputs, clear stale sessions, classify the URL.
-
-        Returns the normalized inputs plus content_type and is_mp4.
-        """
-        _validate_url(remote_url)
-        auth_header = _validate_auth_header(auth_header)
-        fallback_sources = _normalize_fallback_sources(fallback_sources)
-        content_length_hint = _normalize_content_length_hint(content_length_hint)
-        settings_snapshot = normalize_settings_snapshot(settings_snapshot)
-        # Tear down any previous session before starting a new one. Kodi only
-        # ever plays one stream at a time, so anything still in the table is
-        # garbage from a prior play — possibly with a zombie ffmpeg attached
-        # to a half-dead socket if Kodi stalled without firing
-        # onPlayBackStopped. Cleaning up here guarantees the next play gets a
-        # fresh proxy state and no stale ffmpeg hogging the upstream.
-        self.clear_sessions(wait_for_process=False)
-        content_type = self._detect_content_type(remote_url)
-        lower_url = remote_url.lower()
-        is_mp4 = lower_url.endswith((".mp4", ".m4v"))
-        return (
-            auth_header,
-            fallback_sources,
-            content_length_hint,
-            settings_snapshot,
-            content_type,
-            is_mp4,
-        )
-
-    def _finalize_and_register(self, ctx, fallback_sources, settings_snapshot, started):
-        """Attach runtime fields, start prefetch, register session, return info."""
-        _attach_fallback_context_fields(ctx, fallback_sources)
-        if settings_snapshot:
-            ctx[_PASSTHROUGH_RUNTIME_SETTINGS_KEY] = (
-                _passthrough_runtime_settings_from_snapshot(settings_snapshot)
-            )
-        self._start_passthrough_runtime_settings_prefetch(ctx)
-        self._start_initial_range_prefetch(ctx)
-        self._start_readahead_prefetch(ctx)
-        self._start_tail_prewarm(ctx)
-        self._start_fallback_prevalidation(ctx)
-
-        local_url = self._register_session(ctx)
-        xbmc.log(
-            "NZB-DAV: Proxy ready (remux={}, faststart={}): {}".format(
-                ctx.get("remux", False), ctx.get("faststart", False), local_url
-            ),
-            xbmc.LOGINFO,
-        )
-        stream_info = {
-            "duration_seconds": ctx.get("duration_seconds"),
-            "total_bytes": ctx.get("total_bytes", ctx.get("content_length", 0)),
-            "virtual_size": ctx.get("virtual_size", 0),
-            "seekable": (
-                ctx.get("seekable", False)
-                or ctx.get("faststart", False)
-                or ctx.get("temp_faststart", False)
-            ),
-            "remux": ctx.get("remux", False),
-            "faststart": ctx.get("faststart", False),
-            "direct": False,
-            "mode": ctx.get("mode"),
-            "content_type": ctx.get("content_type"),
-        }
-        _attach_fallback_context_fields(
-            stream_info, ctx.get("fallback_sources", fallback_sources)
-        )
-        telemetry.log_timing(
-            "prepare_stream",
-            (time.monotonic() - started) * 1000.0,
-            content_type=ctx.get("content_type"),
-            faststart=ctx.get("faststart", False),
-            remux=ctx.get("remux", False),
-        )
-        return local_url, stream_info
-
-    def prepare_stream(
-        self,
-        remote_url,
-        auth_header=None,
-        fallback_sources=None,
-        content_length_hint=None,
-        settings_snapshot=None,
-    ):
-        """Set up proxy for a new stream.
-
-        Returns (local_proxy_url, stream_info_dict).
-        stream_info_dict contains duration_seconds, total_bytes, seekable, remux,
-        faststart, and virtual_size.
-        """
-        started = time.monotonic()
-        (
-            auth_header,
-            fallback_sources,
-            content_length_hint,
-            settings_snapshot,
-            content_type,
-            is_mp4,
-        ) = self._validate_and_classify(
-            remote_url,
-            auth_header,
-            fallback_sources,
-            content_length_hint,
-            settings_snapshot,
-        )
-
-        ctx = self._build_stream_context(
-            remote_url,
-            auth_header,
-            content_length_hint,
-            content_type,
-            is_mp4,
-            fallback_sources,
-            settings_snapshot,
-        )
-        return self._finalize_and_register(
-            ctx, fallback_sources, settings_snapshot, started
-        )
-
-    def _build_stream_context(
-        self,
-        remote_url,
-        auth_header,
-        content_length_hint,
-        content_type,
-        is_mp4,
-        fallback_sources,
-        settings_snapshot,
-    ):
-        """Decide playback mode and build the matching stream context."""
-        if fallback_sources:
-            return self._build_ctx_fallback(
-                remote_url, auth_header, content_length_hint, content_type, is_mp4
-            )
-        if is_mp4:
-            return self._build_ctx_mp4(remote_url, auth_header, content_length_hint)
-        return self._build_ctx_default(
-            remote_url,
-            auth_header,
-            content_length_hint,
-            content_type,
-            settings_snapshot,
-        )
-
-    @staticmethod
-    def _probe_duration(ffmpeg_path, url, auth_header):
-        """Probe file duration. Returns seconds or None.
-
-        Two strategies, tried in order:
-
-        1. ``ffprobe -show_entries format=duration`` — the clean path. One
-           number on stdout, no stream-probe warnings. This is the only
-           reliable approach for files with many subtitle streams: a 30-
-           subtitle Blu-ray remux produces a wall of ``Could not find
-           codec parameters for stream N (Subtitle: hdmv_pgs_subtitle)``
-           warnings from ffmpeg that can trivially push ``Duration:`` past
-           any stderr buffer budget before it gets a chance to print.
-        2. ``ffmpeg -i <url> -f null -`` parsed out of stderr — the
-           fallback path when ffprobe isn't installed. Budget is 64 KB
-           (up from the original 8 KB) so the subtitle-warning wall
-           doesn't evict the Duration line on pathological inputs.
-
-        Args:
-            ffmpeg_path: Path to the ffmpeg binary. Used by the fallback
-                path and as the starting point for ffprobe discovery.
-            url: Remote HTTP URL to probe.
-            auth_header: Optional Basic auth header; passed to the child
-                process via ``-headers`` so the input URL stays clean.
-        """
-        _validate_url(url)
-        auth_args = _ffmpeg_auth_args(auth_header)
-
-        ffprobe_path = _find_ffprobe()
-        if ffprobe_path:
-            result = StreamProxy._probe_duration_ffprobe(
-                ffprobe_path, url, auth_args=auth_args
-            )
-            if result is not None:
-                return result
-
-        return StreamProxy._probe_duration_ffmpeg(ffmpeg_path, url, auth_args=auth_args)
-
-    @staticmethod
-    def _probe_duration_ffprobe(ffprobe_path, input_url, auth_args=None):
-        """Run ffprobe to get duration. Returns seconds or None."""
-        try:
-            cmd = [
-                ffprobe_path,
-                "-v",
-                "error",
-            ]
-            if auth_args:
-                cmd.extend(auth_args)
-            cmd.extend(
-                [
-                    "-show_entries",
-                    "format=duration",
-                    "-of",
-                    "default=nokey=1:noprint_wrappers=1",
-                    input_url,
-                ]
-            )
-            proc = subprocess.Popen(  # nosec B603 — argv list, shell=False
-                cmd,
-                stdin=subprocess.DEVNULL,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-                shell=False,
-            )
-            try:
-                stdout_bytes, _ = proc.communicate(timeout=30)
-            except subprocess.TimeoutExpired:
-                proc.kill()
-                try:
-                    proc.communicate(timeout=5)
-                except (OSError, subprocess.SubprocessError):
-                    pass
-                xbmc.log("NZB-DAV: ffprobe duration timed out", xbmc.LOGWARNING)
-                return None
-            if proc.returncode != 0:
-                return None
-            text = stdout_bytes.decode("utf-8", "replace").strip()
-            if not text:
-                return None
-            try:
-                return float(text)
-            except ValueError:
-                return None
-        except (OSError, subprocess.SubprocessError) as e:
-            xbmc.log("NZB-DAV: ffprobe failed: {}".format(e), xbmc.LOGWARNING)
-            return None
-
-    @staticmethod
-    def _probe_duration_ffmpeg(ffmpeg_path, input_url, auth_args=None):
-        """Parse Duration out of ``ffmpeg -i`` stderr. Returns seconds or None.
-
-        Uses the bounded-reader-thread pattern: a daemon thread reads
-        stderr line-by-line into a shared buffer and signals an Event
-        as soon as ``Duration:`` is matched or the 64 KB byte budget
-        is exhausted. The main thread waits on the Event with a
-        hard wall-clock deadline of ``_PROBE_DEADLINE_SECONDS`` so a
-        stuck ffmpeg (slow upstream, stalled header parse, auth hang)
-        can't wedge the probe forever. Either way — match, budget,
-        deadline — the ffmpeg process is killed before returning.
-        """
-        return StreamProxy._probe_ffmpeg_stderr(
-            ffmpeg_path,
-            input_url,
-            _parse_ffmpeg_duration,
-            "Duration",
-            auth_args=auth_args,
-        )
-
-    @staticmethod
-    def _probe_ffmpeg_stderr(ffmpeg_path, input_url, parser, label, auth_args=None):
-        """Shared body of ``_probe_duration_ffmpeg`` (DV probing now uses
-        the pure-Python ``dv_source.probe_dolby_vision_source``, which
-        parses real RPU data instead of relying on ffmpeg's stderr).
-        Spawns ``ffmpeg -v info -i <url> -f null -`` and runs the parser
-        against collected stderr under a bounded reader thread +
-        wall-clock deadline.
-
-        ``auth_args`` is an optional list of ffmpeg argv pieces (from
-        ``_ffmpeg_auth_args``) that gets inserted before ``-i`` so the
-        Authorization header is passed via ``-headers`` instead of
-        being spliced into the input URL. Callers that build the URL
-        with ``_embed_auth_in_url`` should leave this None; new
-        callers should prefer the ``-headers`` form.
-        """
-        cmd = [ffmpeg_path, "-v", "info"]
-        if auth_args:
-            cmd.extend(auth_args)
-        cmd.extend(["-i", input_url, "-f", "null", "-"])
-
-        proc = StreamProxy._spawn_probe_proc(cmd, label)
-        if proc is None:
-            return None
-
-        collected = [""]
-        done = threading.Event()
-        # 64 KB budget: large enough that a 30-subtitle Blu-ray remux's
-        # wall of per-stream probe warnings can't push the match line out.
-        budget = 65536
-
-        reader = threading.Thread(
-            target=StreamProxy._probe_stderr_reader,
-            args=(proc, parser, collected, budget, done, label),
-            name="nzbdav-probe-reader",
-            daemon=True,
-        )
-        reader.start()
-
-        if not done.wait(timeout=_PROBE_DEADLINE_SECONDS):
-            xbmc.log(
-                "NZB-DAV: {} probe wall-clock deadline ({}s) exceeded, "
-                "killing ffmpeg".format(label, _PROBE_DEADLINE_SECONDS),
-                xbmc.LOGWARNING,
-            )
-
-        return StreamProxy._finish_probe(proc, reader, parser, collected, budget, label)
-
-    @staticmethod
-    def _spawn_probe_proc(cmd, label):
-        """Spawn an ffmpeg probe process; log + return None on spawn failure."""
-        try:
-            return subprocess.Popen(  # nosec B603 — argv list, shell=False
-                cmd,
-                stdin=subprocess.DEVNULL,
-                stdout=subprocess.DEVNULL,
-                stderr=subprocess.PIPE,
-                shell=False,
-            )
-        except (OSError, subprocess.SubprocessError, ValueError) as e:
-            xbmc.log(
-                "NZB-DAV: {} probe spawn failed: {}".format(label, e),
-                xbmc.LOGWARNING,
-            )
-            return None
-
-    @staticmethod
-    def _probe_stderr_reader(proc, parser, collected, budget, done, label):
-        """Read ffmpeg stderr into ``collected`` until match/budget/EOF."""
-        try:
-            for line in proc.stderr:
-                collected[0] += line.decode(errors="replace")
-                if parser(collected[0]) is not None:
-                    return
-                if len(collected[0]) > budget:
-                    return
-        except Exception as exc:  # pylint: disable=broad-except
-            # Log the failure mode rather than swallowing silently —
-            # a stderr decode error or pipe close that hides a real
-            # probe failure used to surface as duration=None with no
-            # diagnostic in kodi.log. Closes TODO.md §H.3 silent
-            # stderr-reader failure.
-            xbmc.log(
-                "NZB-DAV: {} probe stderr-reader failed: {}".format(label, exc),
-                xbmc.LOGWARNING,
-            )
-        finally:
-            done.set()
-
-    @staticmethod
-    def _finish_probe(proc, reader, parser, collected, budget, label):
-        """Kill the probe proc, join its reader, and return the parsed result."""
-        try:
-            proc.kill()
-        except OSError:
-            pass
-        try:
-            proc.wait(timeout=5)
-        except (subprocess.TimeoutExpired, OSError):
-            pass
-
-        # Join the reader thread now that the proc is dead — its EOF on
-        # stderr is the loop-exit condition, so once kill() takes effect
-        # the thread should terminate within milliseconds. Joining here
-        # (with a small timeout safety net) prevents probe threads from
-        # accumulating in long-running services that probe a lot, e.g. a
-        # search session that prepares many candidate streams.
-        # daemon=True still covers the pathological case where stderr
-        # never EOFs. Closes TODO.md §H.3 probe-reader thread leak.
-        reader.join(timeout=2)
-
-        result = parser(collected[0])
-        if result is None and len(collected[0]) > budget:
-            xbmc.log(
-                "NZB-DAV: {} not found in first {}B of ffmpeg output".format(
-                    label, budget
-                ),
-                xbmc.LOGWARNING,
-            )
-        return result
-
-    @staticmethod
-    def _prepare_tempfile_faststart(ffmpeg_path, url, auth_header):
-        """Remux MP4 with faststart to a temp file. Returns path or None."""
-        if not ffmpeg_path:
-            return None
-
-        _validate_url(url)
-        fd, temp_path = tempfile.mkstemp(
-            prefix="nzbdav_faststart_",
-            suffix=".mp4",
-        )
-        os.close(fd)
-
-        cmd = StreamProxy._build_faststart_cmd(ffmpeg_path, url, auth_header, temp_path)
-        return StreamProxy._run_faststart_remux(cmd, temp_path)
-
-    @staticmethod
-    def _build_faststart_cmd(ffmpeg_path, url, auth_header, temp_path):
-        """Build the temp-file faststart remux ffmpeg argv."""
-        auth_args = _ffmpeg_auth_args(auth_header)
-        cmd = [
-            ffmpeg_path,
-            "-v",
-            "warning",
-            "-y",
-            "-reconnect",
-            "1",
-            "-reconnect_streamed",
-            "1",
-        ]
-        if auth_args:
-            cmd.extend(auth_args)
-        cmd.extend(
-            [
-                "-i",
-                url,
-                "-map",
-                "0",
-                "-c",
-                "copy",
-                "-movflags",
-                "+faststart",
-                temp_path,
-            ]
-        )
-        return cmd
-
-    @staticmethod
-    def _run_faststart_remux(cmd, temp_path):
-        """Run the faststart remux, returning temp_path on success.
-
-        On any failure (non-zero exit, timeout, subprocess error) the
-        partial output is removed and None is returned.
-        """
-        proc = None
-        try:
-            xbmc.log("NZB-DAV: Temp-file faststart remux starting", xbmc.LOGINFO)
-            proc = subprocess.Popen(  # nosec B603 — argv list, shell=False
-                cmd,
-                stdin=subprocess.DEVNULL,
-                stdout=subprocess.PIPE,
-                stderr=subprocess.PIPE,
-                shell=False,
-            )
-            _, stderr = proc.communicate(timeout=600)  # 10 min timeout
-            result = StreamProxy._faststart_remux_result(proc, stderr, temp_path)
-            if result is not None:
-                return result
-            # Fall through to the cleanup block below instead of
-            # returning early — the partial output that ffmpeg
-            # leaves on a failed remux otherwise leaks until the OS
-            # next clears tempdir. Closes TODO.md §H.3 (mkstemp
-            # leak on TimeoutExpired / SubprocessError) for the
-            # ffmpeg-non-zero-exit case in particular.
-        except subprocess.TimeoutExpired as e:
-            # communicate() timing out does NOT kill the child. Without
-            # an explicit kill + reap, the ffmpeg process orphans and
-            # holds the output fd + the inbound HTTP socket, potentially
-            # for hours. Kill + drain the pipe before the exception
-            # propagates; .communicate() on the killed proc reaps it.
-            xbmc.log(
-                "NZB-DAV: Temp faststart timed out after 600s; killing ffmpeg "
-                "(reason=temp_faststart_timeout)",
-                xbmc.LOGWARNING,
-            )
-            StreamProxy._kill_and_reap_faststart(proc)
-            _ = e  # keep linters quiet; exception detail already logged
-        except (OSError, subprocess.SubprocessError) as e:
-            xbmc.log("NZB-DAV: Temp faststart error: {}".format(e), xbmc.LOGWARNING)
-            # Non-timeout subprocess errors usually mean Popen itself
-            # failed or communicate() hit a pipe error. Still try to
-            # reap the child defensively — it's cheap when the proc
-            # already exited and essential when it didn't.
-            if proc is not None and proc.poll() is None:
-                StreamProxy._kill_and_reap_faststart(proc)
-        StreamProxy._remove_temp_file(temp_path)
-        return None
-
-    @staticmethod
-    def _faststart_remux_result(proc, stderr, temp_path):
-        """Evaluate a completed faststart proc.
-
-        Returns temp_path on success (rc==0 and non-empty output), or
-        None on failure (after logging the redacted ffmpeg stderr).
-        """
-        if proc.returncode != 0:
-            # ffmpeg error messages routinely echo the full input URL,
-            # including embedded basic-auth (legacy callers) and
-            # apikey=... query strings. Strip those before they land in
-            # kodi.log. Closes TODO.md §H.2-H2b.
-            xbmc.log(
-                "NZB-DAV: Temp faststart failed: {}".format(
-                    _redact_text(stderr.decode(errors="replace")[:300])
-                ),
-                xbmc.LOGWARNING,
-            )
-            return None
-        if os.path.exists(temp_path) and os.path.getsize(temp_path) > 0:
-            return temp_path
-        return None
-
-    @staticmethod
-    def _kill_and_reap_faststart(proc):
-        """Kill + drain a faststart ffmpeg child (documented reap idiom)."""
-        try:
-            proc.kill()
-            proc.communicate(timeout=5)
-        except (OSError, subprocess.SubprocessError):
-            pass
-
-    @staticmethod
-    def _remove_temp_file(temp_path):
-        """Best-effort removal of a temp file if it exists."""
-        if os.path.exists(temp_path):
-            try:
-                os.remove(temp_path)
-            except OSError:
-                pass
-
-    @staticmethod
-    def _get_content_length(url, auth_header, content_length_hint=None):
-        """Get file size via HEAD or range probe."""
-        content_length_hint = _normalize_content_length_hint(content_length_hint)
-        if content_length_hint > 0:
-            confirmed = _probe_content_length_hint(
-                url, auth_header, content_length_hint
-            )
-            if confirmed > 0:
-                return confirmed
-
-        req = Request(url, method="HEAD")
-        _add_request_headers(req, auth_header)
-        try:
-            # nosemgrep
-            with urlopen(  # nosec B310 — URL from user-configured nzbdav/WebDAV setting
-                req, timeout=10
-            ) as resp:
-                return int(resp.headers.get("Content-Length", 0))
-        except (OSError, ValueError):
-            pass
-        return _probe_content_length_tail(url, auth_header)
-
-    @staticmethod
-    def _detect_content_type(url):
-        """Detect content type from URL extension."""
-        lower = url.lower()
-        if lower.endswith(".mkv"):
-            return "video/x-matroska"
-        if lower.endswith((".mp4", ".m4v")):
-            return "video/mp4"
-        if lower.endswith(".avi"):
-            return "video/x-msvideo"
-        return "video/mp4"
 
 
 def get_service_proxy_port():
